+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-02T08:26:14.665Z",
+    "2026-09-03T07:45:51.499Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {