+++ packages/orq-rc/RELEASES.md
@@ -12049,3 +12049,13 @@ Based on:
 - [typescript v4.15.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.15 - packages/orq-rc
+
+## 2026-09-03 07:48:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.3 (2.934.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.16 - packages/orq-rc